Luke: moves Coverage Model manipulation into ingst
CoverageCraft was the utility responsible for making changes to the
coverage model instance. This utility is no longer beneficial to keep
outside in a separate module so I'm moving the utility of it into the
appropriate DM processes (ingest/replay).  This will ensure that the
interface to manipulate coverage models are kept inside DM and not
exposed outwardly until necessary to do so.