Luke: Fuzzy searching and SN improvements
Term searching now matches exact values, no more hurtles with analyzers.
This means we can search against emails exactly, and any fields which
contain special characters (including non-base64 identifiers).

Wildcard searches are unchanged, if the query has a '*' in the search
string it will promote the query to a wildcard search, which is NOT
analyzed by ES (more strict on matching).

Fuzzy searches are now implemented.  To use a fuzzy search against a
field use the 'LIKE' keyword or if the client is using query_term
directly, set the fuzzy keyword argument to True.  Please see the
integration tests for examples.