Luke: Patches test_helper
If the queue_name in "workflow_test" is being reused there are certain
conditions which could cause leftover messages in the queue from being
cleaned up.  This purges and deletes the queue at the end of the test.
If this is the source of the failures for the
test_SA_transform_components then we should stop seeing this failure.