Removes queue memoization from ingestion
There is a very  unique and strange bug inside of Python's C layer which
causes the container to go into an infinite loop when I assign a tuple
consisting of a RLock and a Queue into an OrderedDict and then return
the references to the objects. I can not trick or in any other feasible
manner provide memoization for the internal queues, therefore they will
be plain and not memoized in any way.