Fixes bool issue with preload.
The workflow to process the preload spreadsheets are divergent between
online and offline resources. If the preload is done using online
resources an XLS parser parses an XLS document; whereby it replaces the
cell values for TRUE and FALSE with 1 and 0 respectively, whereas the
CSV parser used for offline processing leaves the cells in their
original states.

These changes provide a simple _is_true() method for checking known
acceptable boolean true vales in the spreadsheets.