Fixes bug in create_data_process
To support backwards compatibility we supported the bindings of older
transforms. The newer transforms just use the stream ids that way there
doesn't have to be an explicit mapping from a named attribute to a
stream's publisher.  There is a bug where the effort of making the
stream publisher is duplicated, this patch makes the actions exclusive
to one-another.