Adds checksum verification to the downloaded eggs
There is a certain situation where a partially downloaded or corrupted
egg file will remain in the filesystem and a container will fail to load
without any apparent reason other than an ImportError "Can't load
mi.core...".  The source of this problem stems from a corrupted egg.
This checksum comparison at least provides a more detailed message as to
what's going on.

The solution to get the container to load again is to delete the egg in
'/tmp' and try again. MI may want to add some special handling here to
try the download again or something.