Fixes two data product related tests
register_dataset no longer takes advantage of the registered attribute
for a dataset resource, so the test no longer needs to verify that.

activate_data_product on a derived data product is now automatic during
creation. Calling activate_data_product on a derived data product
results in a BadRequest being raised, so this test got patched.