Adds a configurable timeout to dataset monitor based on global config
Timeouts on the buildbot are a frequent occurrence for the dataset
monitor when we use small static values. This commit updates all
instances of this pattern that I could find and switches it to use a
dynamic configurable parameter through pyon.yml

Right now it uses endpoint.receive.timeout but we may want to use a
different CFG key but at least now the failures on buildbot should be
redeuced in frequency.